CLI: Expose sign-only reply parsing helper
